Modifications to work with MS Windows
lxml only works with posix type paths so change path to this type
to start with by the function posix_path().
also normpath removed in EpubWriter.py.
This works with the test file but needs more comprehensive
testing.
Since Python works  on windows with posix paths it might be
possible to remove all os functions